refactor: move types into their own respective files
