+++ src/types/common.ts
@@ -1,0 +1,27 @@
+/*
+
+
+
+
+
+
+
+
+*/
+
+/*
+ * Basic theme object shape
+ *
+ * @since 1.0.0
+ */
+export type BaseExtensibleObject = {
+  [key: string]: any
+}
+
+/**
+ * Basic theme object shape
+ *
+ * @since 1.0.0
+ */
+// eslint-disable-next-line
+export type CSSProperty = string | number | (string & {}) | (number & {})
